BackEnd/django/requirements.txt
django==1.10.3
djangorestframework==3.5.3
django-cors-headers==1.1.0